[#9] Move constants
