--- src/constants.js
@@ -1,26 +1,0 @@
-export const MAX_RESULTS = 100;
-export const MAX_HISTORY_RESULTS = 100;
-export const MAX_BOOKMARK_RESULTS = 1000;
-
-export const ItemTypes = {
-  TAB: 'tab',
-  HISTORY: 'history',
-  BOOKMARK: 'bookmark'
-};
-
-export const Commands = {
-  TOGGLE_ANCHOR: 'toggle-anchor',
-  TOGGLE_ANCHOR_WITH_BOOKMARK_MODE: 'toggle-anchor-with-bookmark-mode',
-  TOGGLE_ANCHOR_WITH_HISTORY_MODE: 'toggle-anchor-with-history-mode'
-};
-
-export const KeyCodes = {
-  RETURN: 13,
-  ESCAPE: 27,
-  LEFT: 37,
-  UP: 38,
-  RIGHT: 39,
-  DOWN: 40,
-  N: 78,
-  P: 80
-};
