+++ src/constants/commands.js
@@ -1,0 +1,3 @@
+export const TOGGLE_ANCHOR = 'toggle-anchor';
+export const TOGGLE_ANCHOR_WITH_BOOKMARK_MODE = 'toggle-anchor-with-bookmark-mode';
+export const TOGGLE_ANCHOR_WITH_HISTORY_MODE = 'toggle-anchor-with-history-mode';
