+++ src/constants/item_types.js
@@ -1,0 +1,3 @@
+export const TAB = 'tab';
+export const HISTORY = 'history';
+export const BOOKMARK = 'bookmark';
