+++ src/constants/key_codes.js
@@ -1,0 +1,8 @@
+export const RETURN = 13;
+export const ESCAPE = 27;
+export const LEFT = 37;
+export const UP = 38;
+export const RIGHT = 39;
+export const DOWN = 40;
+export const N = 78;
+export const P = 80;
