+++ src/utils/bookmark_manager.js
@@ -1,5 +1,7 @@
 import EventEmitter from 'events';
-import { MAX_BOOKMARK_RESULTS, ItemTypes } from '../constants';
+import * as ItemTypes from '../constants/item_types';
+
+const MAX_BOOKMARK_RESULTS = 1000;
 
 export default class BookmarkManager extends EventEmitter {
   constructor() {
