+++ src/utils/tab_manager.js
@@ -1,5 +1,5 @@
 import EventEmitter from 'events';
-import { ItemTypes } from '../constants';
+import * as ItemTypes from '../constants/item_types';
 
 export default class TabManager extends EventEmitter {
   constructor() {
